pochti

+++ generator.h
@@ -1,4 +1,5 @@
 #include <string>
+#include <cstdint>
 using namespace std;
 
-string autoGenerateKey(const string& cipherName, size_t textLength);
+string autoGenerateKey(const string& cipherName, uint64_t textLength);
