+++ utils.cpp
@@ -5,13 +5,10 @@
 #include <iomanip>
 #include <fstream>
 #include <sstream>
-
+#include <cstdint>
 using namespace std;
 
-
-
-
-// Чтение бинарного файла
+// Чтение бинарника
 string readBinaryFile(const string& filename) {
     ifstream file(filename, ios::binary);
     if (!file) {
@@ -23,7 +20,7 @@ string readBinaryFile(const string& filename) {
     return buffer.str();
 }
 
-// Запись бинарного файла
+// запись бинарника
 bool writeBinaryFile(const string& filename, const string& data) {
     ofstream file(filename, ios::binary);
     if (!file) {
@@ -59,7 +56,7 @@ string fromHex(const string& hex) {
     string result;
     if (hex.size() % 2 != 0) return "";
 
-    for (size_t i = 0; i < hex.size(); i += 2) {
+    for (uint64_t i = 0; i < hex.size(); i += 2) {
         unsigned int byte;
         if (sscanf(hex.substr(i, 2).c_str(), "%02X", &byte) != 1)
             return "";
